Fix #2039855 [Exception when opening LRF file - QFont.setPixelSize gets float instead of int](https://bugs.launchpad.net/calibre/+bug/2039855)

+++ src/calibre/gui2/lrf_renderer/text.py
@@ -401,7 +401,7 @@ class Line(QGraphicsItem):
         matches = self.__class__.whitespace.finditer(phrase)
         font = QFont(ts.font)
         if self.valign is not None:
-            font.setPixelSize(font.pixelSize()/1.5)
+            font.setPixelSize(int(font.pixelSize()/1.5))
         fm = QFontMetrics(font)
         single_space_width = fm.horizontalAdvance(' ')
         height, descent = fm.height(), fm.descent()
